Save timestamped recording
So that we know when in time the recording took place.
Keeping meta data in filename might not be enough though.

Also the URL for the upstream is not the same as the
resource created. Some mechanism either sending message back
via websocket or redirecting GET to the correct resource
must be implemented.